Updated logger settings, query databases working
