+++ src/aerospace_chatbot/core/__init__.py
@@ -7,9 +7,10 @@ from .config import (
     ConfigurationError,
     load_config,
     get_secrets,
-    set_secrets
+    set_secrets,
+    setup_logging
 )
-from .logging_config import setup_logging
+
 __all__ = [
     'Dependencies',
     'ConfigurationError',
