--- src/aerospace_chatbot/core/logging_config.py
@@ -1,51 +1,0 @@
-import logging
-import sys
-from pathlib import Path
-import os
-import warnings
-
-def setup_logging():
-    """Configure logging based on environment variables.
-    
-    Environment Variables:
-        LOG_LEVEL: DEBUG, INFO, WARNING, ERROR, CRITICAL (default: INFO)
-        LOG_FILE: Path to log file (optional)
-        LOG_FORMAT: Custom log format (optional)
-    """
-    # Suppress Streamlit warning about missing ScriptRunContext
-    warnings.filterwarnings('ignore', message='.*missing ScriptRunContext.*')
-    
-    # Get settings from environment
-    log_level = os.getenv('LOG_LEVEL', 'INFO').upper()
-    log_file = os.getenv('LOG_FILE')
-    log_format = os.getenv('LOG_FORMAT', 
-        '%(asctime)s - %(name)s - %(levelname)s - [%(filename)s:%(lineno)d] - %(message)s'
-    )
-
-    # Convert string log level to logging constant
-    numeric_level = getattr(logging, log_level, None)
-    if not isinstance(numeric_level, int):
-        raise ValueError(f'Invalid log level: {log_level}')
-
-    # Configure root logger
-    root_logger = logging.getLogger()
-    root_logger.setLevel(numeric_level)
-
-    # Create formatter
-    formatter = logging.Formatter(log_format)
-
-    # Always set up console handler
-    console_handler = logging.StreamHandler(sys.stdout)
-    console_handler.setFormatter(formatter)
-    root_logger.addHandler(console_handler)
-
-    # Optionally set up file handler
-    if log_file:
-        log_path = Path(log_file)
-        log_path.parent.mkdir(parents=True, exist_ok=True)
-        
-        file_handler = logging.FileHandler(log_file)
-        file_handler.setFormatter(formatter)
-        root_logger.addHandler(file_handler)
-
-    return root_logger
